Moved create accomplishment and skill operations to POST /api/resume/[candidateId]/experience/[roleId]/accomplishments and POST /api/resume/[candidateId]/skillSets/[skillSetId]/skills, respectively.
